+++ sample-new-architecture/metro.config.js
@@ -2,6 +2,9 @@ const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config');
 const path = require('path');
 const blacklist = require('metro-config/src/defaults/exclusionList');
 
+const {
+  createSentryMetroSerializer,
+} = require('../dist/js/tools/sentryMetroSerializer');
 const parentDir = path.resolve(__dirname, '..');
 
 /**
@@ -41,6 +44,10 @@ const config = {
       },
     ),
   },
+  serializer: {
+    customSerializer: createSentryMetroSerializer(),
+  },
 };
 
-module.exports = mergeConfig(getDefaultConfig(__dirname), config);
+const m = mergeConfig(getDefaultConfig(__dirname), config);
+module.exports = m;